speckle-build.sh: Copy reference outputs to the target's output-reference
This is quite handy to have an opportunity to compare generated and expected
results. One caveat though is reference outputs have funky names referring to
input files and we on execution name outputs after the benchmark name + ".index".

So direct comparison with diff utility won't work, you'd need to compare
outputs manually. Fortunately except for "perlbench" amount of outputs
is not that huge, so perfectly doable by hand & eye.