Revert "speckle-run.sh: Allow specify amount of instances executed in parallel"
This reverts commit d26baaebd76795d7c54874840c6c11958ad24352.

Proposed implementation was incorrect because instead of parallel
execution of complete packages (yep, quite some benchmarks consist
of multiple executions of the same binary but with different input data)
we parallel execution of each separate command line. Effectively spoiling
execution of even SPECspeed (no parallel execution of benchmarks).

So to support true SPECrate we need to move execution of the entire
benchmark package (no matter there will be 1 binary execution or
many of those) to the separate function and then have multiple
instances of that function executed.